+++ NEWS.rst
@@ -9,4 +9,6 @@ Latest
 * Major: Added level member to metrics class
 * Minor: Added read_storage() to metrics class
 * Major: Separated the metric class from metrics
+* Major: Added class view to handle the raw memory of class metrics
+* Minor: Added class view_iterator